fix #54 Corrects some methods in Parameters, which had the wrong return type.

+++ src/Parameters.java
@@ -33,23 +33,23 @@ public class Parameters implements java.io.Serializable{
     public void setDIST(double DIST){this.DIST=DIST;}
     public double getDIST(){return DIST;}
     public void setN_PTS(int N_PTS){this.N_PTS=N_PTS;}
-    public double getN_PTS(){return N_PTS;}
+    public int getN_PTS(){return N_PTS;}
     public void setK_PTS(int K_PTS){this.K_PTS=K_PTS;}
-    public double getK_PTS(){return K_PTS;}
+    public int getK_PTS(){return K_PTS;}
     public void setA_PTS(int A_PTS){this.A_PTS=A_PTS;}
-    public double getA_PTS(){return A_PTS;}
+    public int getA_PTS(){return A_PTS;}
     public void setB_PTS(int B_PTS){this.B_PTS=B_PTS;}
-    public double getB_PTS(){return B_PTS;}
+    public int getB_PTS(){return B_PTS;}
     public void setC_PTS(int C_PTS){this.C_PTS=C_PTS;}
-    public double getC_PTS(){return C_PTS;}
+    public int getC_PTS(){return C_PTS;}
     public void setD_PTS(int D_PTS){this.D_PTS=D_PTS;}
-    public double getD_PTS(){return D_PTS;}
+    public int getD_PTS(){return D_PTS;}
     public void setE_PTS(int E_PTS){this.E_PTS=E_PTS;}
-    public double getE_PTS(){return E_PTS;}
+    public int getE_PTS(){return E_PTS;}
     public void setF_PTS(int F_PTS){this.F_PTS=F_PTS;}
-    public double getF_PTS(){return F_PTS;}
+    public int getF_PTS(){return F_PTS;}
     public void setG_PTS(int G_PTS){this.G_PTS=G_PTS;}
-    public double getG_PTS(){return G_PTS;}
+    public int getG_PTS(){return G_PTS;}
     public void setLength2(double LENGTH2){this.LENGTH2=LENGTH2;}
     public double getLength2(){return LENGTH2;}
     public void setRadius2(double RADIUS2){this.RADIUS2=RADIUS2;}
